Order

+++ Project/Loothub/src/main/java/com/company/springboot/dao/IUserDao.java
@@ -9,6 +9,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 public interface IUserDao extends UserDetailsService{
     public List<User> listAll();
     public  User save(User user);
-    public User get(Long id);
-    public void delete(Long id);
+    public User get(int id);
+    public void delete(int id);
 }
