+++ Project/Loothub/src/main/java/com/company/springboot/repository/IOrdersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IOrdersRepository extends JpaRepository<Orders,Long>{
+public interface IOrdersRepository extends JpaRepository<Orders,Integer>{
     
     public Orders findByUserId(Integer userId);
 }
